// chat-app-frontend/src/app/services/models/chat-preview-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatPreviewResponse {
  chat_name?: string;
  id?: string;
  is_recipient_online?: boolean;
  last_message?: string;
  last_message_time?: string;
  last_message_type?: number;
  recipient_id?: string;
  sender_id?: string;
  unread_count?: number;
}